+++ tests/testthat.R
@@ -1,5 +1,6 @@
 library(testthat)
 library(vdiffr)
+library(withr)
 library(epinowcast)
 
 test_check("epinowcast")