+++ NEWS.md
@@ -43,7 +43,8 @@ This release is in development and not yet ready for production use.
 - Add documentation for all custom stan functions. See #422 by @seabbs and reviewed by @sbfnk.
 - Changed `enw_preprocess_data()` to use the maximum delay observed in the data by default (before, the default was 20 days). Added a function `check_max_delay()` which allows to obtain coverage statistics for the assumed maximum delay based on the observed data. Enhanced postprocessing functions to accept a different max_delay than used in the model, by adding artificial samples/summaries for not-modeled dates. Further improved documentation and warnings around `max_delay`. See #224 by @adrian-lison and @seabbs and reviewed by @seabbs.
 - Exposed `enw_stan_to_r()` to the user. This function is used for testing and in development to expose `{epinowcast}` stan code in R. Users may find this function useful as it allows them to explore the stan code used in `{epinowcast}` models more easily. Note that this functionality is known to be unstable when `{rstan}` is loaded in the same R session. See #431 by @seabbs and reviewed by @sbfnk.
-- Refactored `extract_sparse_matrix()` to allow us to drop our `{rstan}` dependency. See #431 by @seabbs and reviewed by @sbfnk.
+- Refactored `extract_sparse_matrix()` to allow us to drop our `{rstan}` dependency.  See #431 by @seabbs and reviewed by @sbfnk.
+- Allow for caching Stan models across R sessions to reduce compilation time through the use of the environment variable, `enw_cache_location`, which can be set using the `set_enw_cache()` function. See #407 by @medewitt and @seabbs and reviewed by @sbfnk and @pearsonca.
 
 ## Model